% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aai.R
\docType{package}
\name{aai}
\alias{aai}
\title{aai: Functions, apps, exercises and other R related stuff used in "AI - Aalborg Intelligence"}
\description{
Functions, apps, exercises and other R related stuff used in "AI - Aalborg Intelligence"
    The project (2020 - 2026) is supported by the Novo Nordisk Foundation to develop teaching material
    to be used in the Danish highschools to strengthen the understanding of AI while explaining how
    basic maths is used in the some popular AI methods.
}
