% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classification_test_data}
\alias{classification_test_data}
\title{Fictive data set used to demonstrate some concepts on classification}
\format{
A data frame with 10 rows and 2 variables:
\describe{
  \item{Længde}{Length}
  \item{Vægt}{Weight}
}
}
\usage{
classification_test_data
}
\description{
A data frame containing the length add weight of SOMETHING.
In `classification_train_data` the class is also given.
}
\keyword{datasets}
