% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preamble.R
\name{dt_simple}
\alias{dt_simple}
\title{Simple function for DT output}
\usage{
dt_simple(tab, ...)
}
\arguments{
\item{tab}{The table to format}

\item{...}{Arguments passed to `DT::datatable`}
}
\description{
Simple function for DT output
}
