% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preamble.R
\name{dt_table}
\alias{dt_table}
\title{Short function for DT output}
\usage{
dt_table(tab, ...)
}
\arguments{
\item{tab}{The table to format}

\item{...}{Arguments passed to `DT::datatable`}
}
\description{
Short function for DT output
}
