% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preamble.R
\name{kable_}
\alias{kable_}
\title{Simple function for kable output}
\usage{
kable_(tab, ...)
}
\arguments{
\item{tab}{The table to format}

\item{...}{Arguments passed to `knitr::kable`}
}
\description{
Simple function for kable output
}
