% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{kMD_plot}
\alias{kMD_plot}
\title{Data for plotting a grid based the mean of the K nearest neighbors}
\usage{
kMD_plot(K = 3, .train, response = "Type", grid = 100)
}
\arguments{
\item{K}{Number of neighbors}

\item{.train}{Training data}

\item{response}{Name of the class variable}

\item{grid}{Resolution of the grid - higher values gives finer grid}
}
\description{
Data for plotting a grid based the mean of the K nearest neighbors
}
