% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{kNN.cv}
\alias{kNN.cv}
\title{Actual cross-validation function for kNN.}
\usage{
kNN.cv(K = 3, .train, response = "Type", fold = 10)
}
\arguments{
\item{K}{Vector of nearest neighbor values (the k in kNN)}

\item{.train}{The data to use kNN on}

\item{response}{The variable name of the response}

\item{fold}{The number of folds to use in cross validation}
}
\description{
Actual cross-validation function for kNN.
}
\examples{
data(classification_train_data)
K_LOO <- tibble(K = 1:15,
LOO = kNN.loo(K, .train = classification_train_data)
) \%>\%
rowwise() \%>\%
mutate(CV = list(kNN.cv(K, .train = classification_train_data)))

K_LOO \%>\% ggplot(aes(x = factor(K))) +
geom_boxplot(data = unnest(K_LOO, CV), aes(y = CV)) +
geom_point(aes(y = LOO), colour = "#999999") +
labs(x = "K", y = "Accuracy")
}
