% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{kNN.loo}
\alias{kNN.loo}
\title{Wrapper around `class::knn.cv` which does Leave one Out (LoO)}
\usage{
kNN.loo(K = 3, .train, response = "Type")
}
\arguments{
\item{K}{Number of nearest neighbors to use (can be a vector)}

\item{.train}{Annotated training data that should be classified the test data}

\item{response}{Name of the response variable}
}
\description{
Wrapper around `class::knn.cv` which does Leave one Out (LoO)
}
