% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{kNN}
\alias{kNN}
\title{Wrapper around `class::knn`}
\usage{
kNN(K = 3, .test, .train, response = "Type")
}
\arguments{
\item{K}{Number of nearest neighbors to use}

\item{.test}{Data that should be classified based on the training data}

\item{.train}{Annotated training data that should be classified the test data}

\item{response}{Name of the response variable}
}
\description{
Wrapper around `class::knn`
}
