% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{kNN1}
\alias{kNN1}
\title{Wrapper around `class::knn1`}
\usage{
kNN1(.test, .train, response = "Type")
}
\arguments{
\item{.test}{Data that should be classified based on the training data}

\item{.train}{Annotated training data that should be classified the test data}

\item{response}{Name of the response variable}
}
\description{
Wrapper around `class::knn1`
}
