% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{kNN_plot}
\alias{kNN_plot}
\title{Visualise a kNN trainer}
\usage{
kNN_plot(K = 3, .train, response = "Type", grid = 100)
}
\arguments{
\item{K}{Number of neighbors to use}

\item{.train}{The training data}

\item{response}{The name of the response/class variable}

\item{grid}{The resolution of the grid. Larger numbers gives higher resolution (and slower performance).}
}
\description{
Visualise a kNN trainer
}
\examples{
k <- 3
kNN_plot(.train = classification_train_data, K = k) \%>\%
ggplot() + labs(title = paste("K =", k)) +
geom_rect(aes(xmin = Længde_0, xmax = Længde_1, ymin = Vægt_0, ymax = Vægt_1, fill = Type), alpha = 0.3) +
geom_point(data = train, aes(x = Længde, y = Vægt, colour = Type))
}
