% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{meandist_to_k_nearest}
\alias{meandist_to_k_nearest}
\title{Mean distance to k nearest}
\usage{
meandist_to_k_nearest(
  K = 3,
  .test,
  .train,
  response = "Type",
  dist = FALSE,
  info = TRUE
)
}
\arguments{
\item{K}{Number of nearest neighbors}

\item{.train}{The training data}

\item{return_all}{Logical. Should the distance to the nearest K be returned or just the mean distance of them?}
}
\value{
If `return_all = FALSE` a dataframe of the mean distance to each class of `response` is returned.
    If `return_all = TRUE` a list is returned - `top_K` is as above, `all` contains the closest neighbors from each class.
}
\description{
Mean distance to k nearest
}
\examples{
data(classification_train_data)
meandist_to_k_nearest_(K = 3, .train = classification_train_data) \%>\%
  mutate(same_Type = ifelse(obs_Type == Type, "Y", "N")) \%>\%
  ggplot(aes(x = obs_Type, y = Distance, fill = Type, colour = same_Type)) +
  labs(x = "Type of the observation", fill = "Type of the nearest points") +
  theme(legend.position = "top") +
  guides(colour = FALSE) + scale_colour_manual(values = c("Y" = "#666666", "N" = "#000000")) +
  geom_boxplot() + coord_flip()
}
