% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN.R
\name{meandist_to_k_nearest_}
\alias{meandist_to_k_nearest_}
\title{Mean distance to k nearest}
\usage{
meandist_to_k_nearest_(K = 5, .train, response = "Type", return_all = FALSE)
}
\arguments{
\item{K}{Number of nearest neighbors}

\item{.train}{The training data}

\item{return_all}{Logical. Should the distance to the nearest K be returned or just the mean distance of them?}
}
\value{
If `return_all = FALSE` a dataframe of the mean distance to each class of `response` is returned.
    If `return_all = TRUE` a list is returned - `top_K` is as above, `all` contains the closest neighbors from each class.
}
\description{
Mean distance to k nearest
}
