% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perceptron31}
\alias{perceptron31}
\title{Fictive data set used to demonstrate some concepts in perceptron document}
\format{
A data frame with 31 rows and 3 variables:
\describe{
  \item{x1}{Answer to first question.}
  \item{x2}{Answer to second question.}
  \item{col}{Class}
}
}
\usage{
perceptron31
}
\description{
A data frame containing the responses to two fictive questions on the scale -2,-1,0,1,2 together with a classification color.
}
\keyword{datasets}
