% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grid.R
\name{pred_grid}
\alias{pred_grid}
\title{Helper function for making preditive grid}
\usage{
pred_grid(
  data,
  step = 10,
  response = "Type",
  pred_var = "Prediction",
  center = 0
)
}
\arguments{
\item{data}{Dataset}

\item{step}{Step size in each data variable}

\item{response}{The name of the response variable}

\item{center}{If not through zero, then through `center`}
}
\description{
Helper function for making preditive grid
}
