% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grid.R
\name{pred_max}
\alias{pred_max}
\title{Method for predicting the majority vote or "?" if ties}
\usage{
pred_max(n, x)
}
\arguments{
\item{n}{Counts}

\item{x}{Data vector}
}
\description{
Method for predicting the majority vote or "?" if ties
}
