% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grid.R
\name{pred_plot_grid}
\alias{pred_plot_grid}
\title{Helper function for making preditive grid}
\usage{
pred_plot_grid(pred_grd, pred_var = "Prediction", remove = TRUE)
}
\arguments{
\item{pred_grd}{Output from YY function}

\item{remove}{Is parsed to the `remove` argument of `tidyr::separate()`}
}
\description{
Helper function for making preditive grid
}
