% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grid.R
\name{predict_grid}
\alias{predict_grid}
\title{Create grid for new data}
\usage{
predict_grid(pred_grd, newdata)
}
\arguments{
\item{pred_grd}{Returned from ?}

\item{newdata}{New data to be used in prediction}
}
\description{
Create grid for new data
}
