% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preamble.R
\name{Print}
\alias{Print}
\title{Makes print return all rows in a tibble}
\usage{
Print(...)
}
\arguments{
\item{...}{Arguments passed to `knitr::kable`}
}
\description{
Makes print return all rows in a tibble
}
