% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grid.R
\name{seq_cut}
\alias{seq_cut}
\title{Create discretised version with some pretty labels}
\usage{
seq_cut(x, step, center, breaks = FALSE)
}
\arguments{
\item{x}{Data variable}

\item{step}{Step size}

\item{center}{If not through zero, then through `center`}

\item{breaks}{Logical. Should break labels we returned?}
}
\description{
Create discretised version with some pretty labels
}
\examples{
seq_cut(rnorm(100), step = 2, center = 0, breaks = TRUE)
}
