% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grid.R
\name{seq_zero}
\alias{seq_zero}
\title{Create breaks for `seq_cut`}
\usage{
seq_zero(x, step, center)
}
\arguments{
\item{x}{Data variable}

\item{step}{Step size}

\item{center}{If not through zero, then through `center`}
}
\description{
Create breaks for `seq_cut`
}
\examples{
seq_zero(rnorm(100), step = 2, center = 0)
}
