% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_kNN_plot.R
\name{xy_plot}
\alias{xy_plot}
\title{Plot of data for exercise by Jan B Sørensen on classification}
\usage{
xy_plot(train, x, y, colour, test = NULL, selected = NULL)
}
\arguments{
\item{train}{Training data set}

\item{x, y, colour}{parameters controlling the x and y axis and point colours}

\item{test}{Test data set}

\item{selected}{points to highlight}
}
\description{
Plot of data for exercise by Jan B Sørensen on classification
}
\examples{
data(classification_train_data)
data(classification_test_data)
type_cols <- c("1" = "#E41A1C", "2" = "#377EB8", "3" = "#4DAF4A", "?" = "#444444")
xy_plot(train = classification_train_data, x = Længde, y = Vægt, colour = Type) +
  scale_colour_manual(values = type_cols)
}
